+++ CHANGES.rst
@@ -6,6 +6,11 @@
  Unreleased
 ************
 
+**Added**
+
+- Added command-line option ``--indent-width`` (4 as default) to control
+  spaces per indent level.
+
 ********************
  2.1.1 (2026/06/12)
 ********************